Adding bin/update_index.py script.
This script generates the footers for all files. Rerun the script after
adding new books or chapters in order to update the footer sections
on each page with links to the next and previous chapters.

The script works by recursively parsing "## Index" sections in files,
starting with README.md. The footer is marked with an HTML comment,
`<!--automatically generated footer-->`. Any text after this comment
is destroyed by the script, so all edits should be made above that
point.